Refactor, repackage, clean up and eliminate unused elements.
